+++ src/game/tilemap.cpp
@@ -34,10 +34,10 @@ entt::entity TileMap::at(const int x, const int y) {
 }
 
 // Public function converting x, y tilemap coordinates to screen coordinates
-glm::vec2 TileMap::grid_to_pixel(const int x, const int y) const {
+glm::ivec2 TileMap::grid_to_pixel(const int x, const int y) const {
     int x_offset {x-y};
     int y_offset {y+x};
-    return glm::vec2 {
+    return glm::ivec2 {
         constants::TILEMAP_X_START + (x_offset * constants::TILE_WIDTH_HALF),
         constants::TILEMAP_Y_START + (y_offset * constants::TILE_HEIGHT_HALF)
     };
